+++ src/main/java/com/meritamerica/assignment6/repository/CheckingAccountRepository.java
@@ -3,6 +3,6 @@ package com.meritamerica.assignment6.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.meritamerica.assignment6.models.CheckingAccount;
 
-public interface CheckingAccountRepository extends JpaRepository<CheckingAccount, Long>{
+public interface CheckingAccountRepository extends JpaRepository<CheckingAccount, Integer>{
 
 }
